added id validation to APIs

+++ server/controllers/layoutController.ts
@@ -37,11 +37,20 @@ export const getLayouts = async (req: Request, res: Response) => {
 
 export const getSingleLayout = async (req: Request, res: Response) => {
     const { id } = req.params
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: "no such layout" })
+    }
     const layout = await Layout.findOne({ _id: id })
+    if (!layout) {
+        return res.status(404).json({ error: "no such layout" })
+    }
     return res.status(200).json(layout)
 }
 export const deleteLayout = async (req: Request, res: Response) => {
     const { id } = req.params
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: "no such layout" })
+    }
     const layout = await Layout.findOneAndDelete({ _id: id })
 
     if (!layout) {
